[FIX] Fix bug of start_date stopping computation
The compute method was being interrupted prematurely by the framework in cases with a large number of tasks. To fix this, the logic was moved to the write method. The approach used to obtain all the tasks dependent on a given one is a breadth-first-search traversal done in get_all_tasks_to_be_updated